mix-analyzer/backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
anthropic>=0.40,<1
python-dotenv>=1.0,<2
pydantic>=2.7,<3
